requirements.txt
aiohttp==3.9.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic[email]==2.5.0
PyJWT==2.8.0
bcrypt==4.1.2